pyproject.toml
[project]
name = "ducon-library-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "chromadb>=1.5.0",
    "fastapi>=0.128.0",
    "google-auth>=2.48.0",
    "passlib[bcrypt]>=1.7.4",
    "pathlib>=1.0.1",
    "pillow>=12.1.0",
    "pydantic[email]>=2.12.5",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.2",
    "sqlalchemy[asyncio]>=2.0.47",
    "uvicorn>=0.40.0",
]
